cv_service/requirements.txt
requests
pillow
fastapi
uvicorn
httpx
python-multipart
ultralytics
